pyproject.toml
[project]
name = "ollama-blog-buddy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bs4>=0.0.2",
    "duckdb>=1.1.3",
    "feedparser>=6.0.11",
    "groq>=0.13.1",
    "lancedb>=0.17.0",
    "langchain-community>=0.3.14",
    "langchain-experimental>=0.3.4",
    "langchain-ollama>=0.2.2",
    "loguru>=0.7.3",
    "rich>=13.9.4",
    "streamlit>=1.41.1",
    "tantivy>=0.22.0",
    "tavily-python>=0.5.0",
    "unstructured[all-docs]>=0.16.11",
    "watchdog>=6.0.0",
]
